+++ iChef-WebApi/src/main/java/com/kitchen/iChef/Controller/Model/Request/LogInRequest.java
@@ -1,4 +1,4 @@
-package com.kitchen.iChef.Controller.Request;
+package com.kitchen.iChef.Controller.Model.Request;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,20 +7,12 @@ import lombok.ToString;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
+import javax.validation.constraints.Size;
 
 @Getter
 @Setter
 @ToString
-public class SignUpRequest {
-
-    @NotBlank(message = "The first name of the user is not valid!")
-    private String firstName;
-
-    @NotBlank(message = "The last name of the user is not valid!")
-    private String lastName;
-
-    @NotBlank(message = "The username of the user is not valid!")
-    private String username;
+public class LogInRequest {
 
     @NotNull(message = "The email of the user is not valid!")
     @Pattern(regexp = "^[a-zA-Z]+[a-zA-Z0-9.]*@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*$",
@@ -28,5 +20,6 @@ public class SignUpRequest {
     private String email;
 
     @NotBlank(message = "The password is not valid!")
+    @Size(min = 8)
     private String password;
 }
