+++ iChef-WebApi/src/main/resources/application.properties
@@ -1,12 +1,11 @@
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/theChefs
+spring.datasource.url=jdbc:mysql://localhost:3306/schema
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=1234
 
 spring.thymeleaf.enabled=false
 
 server.servlet.context-path= /kitchen
-spring.profiles.active=dev
 
 spring.servlet.multipart.max-file-size=1024KB
 spring.servlet.multipart.max-request-size=1024KB
